Relocate functions in AdminOrderList container
Some actions and reducers are better kept in the canonical
project directories for their respective locations. Move all the
functions kept in AdminOrderList as a scratchpad in their corresponding
locations.
